requirements.txt
tensorflow==2.12.0
numpy
pillow
opencv-python
flask
werkzeug
